pyproject.toml
[project]
name = "streamlit-lyrebird-mini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "supabase>=2.15.1",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "openai>=1.78.1",
    "sqlalchemy-cockroachdb>=2.0.2",
    "langsmith>=0.3.42",
    "pyperclip>=1.8.2",
    "streamlit-mic-recorder>=0.0.8",
    "langchain-core>=0.3.59",
    "langchain>=0.3.25",
]
